+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Bugfixes, internal refactors, documentation improvements and style changes will
 not be mentioned here, because they do not impact how the package is to be used.
 
 ## 0.4.2
-* Added `truncate(v, i)` similar to `v[1:i]`, but may be more efficient.
+* Added `MemoryViews.truncate(v, i)` similar to `v[1:i]`, but may be more efficient.
 
 ## 0.4.1
 Various fixes and optimizations.